+++ CHANGELOG.md
@@ -9,15 +9,21 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
-- clazzfish-spi module added 
-- clazzfish-spi-git module added to support import/export of statistics using GIT protocol 
+- clazzfish-spi module added
+- clazzfish-spi-git module added to support import/export of statistics using GIT protocol
+
+
+## [2.5.2] (28-Mar-2025)
+
+### Fixed
+- import performance improved
+  ([issue #26](https://github.com/oboehm/ClazzFish/issues/26))
 
 ## [2.5.1] (12-Mar-2025)
 
 ### Fixed
 - lost loaded class statistic during import
   ([issue #25](https://github.com/oboehm/ClazzFish/issues/25))
-
 
 ## [2.5] (11-Mar-2025)